Remove dd_llmobs_ml_app from config, just use env var. Refactor llmobs test fixtures
